Display an unrelated list
I'm displaying a list, but the info in it is unrelated. I just wanted to learn how to do a list. I can't access the contacts as as soon as I try to get contacts it says the plugin isn't installed. This may be because the contacts plugin is deprecated and requires an older version of something in npm that I need to reinstall.